Steel -Metal manufacturing Defect Detection System/requirements.txt
absl-py==2.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.8.1
astunparse==1.6.3
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.3.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
Django==5.1.4
fastapi==0.124.2
filelock==3.16.1
flatbuffers==24.12.23
fonttools==4.55.3
fsspec==2025.12.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.69.0
h11==0.16.0
h5py==3.12.1
huggingface-hub==0.36.0
idna==3.10
jax==0.5.0
jaxlib==0.5.0
Jinja2==3.1.6
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mediapipe==0.10.20
ml_dtypes==0.5.1
mpmath==1.3.0
mysqlclient==2.2.6
namex==0.0.8
networkx==3.6.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
protobuf==4.25.5
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.1
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.3
rich==13.9.4
safetensors==0.7.0
scipy==1.15.1
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
sqlparse==0.5.3
starlette==0.50.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.38.0
virtualenv==20.28.1
Werkzeug==3.1.3
wrapt==1.17.2
